fixed slug page
