+++ src/posts/2022-04-14_third_post.md
@@ -2,5 +2,7 @@
 title: Third Post
 date: 2022-04-14T09:29:41.463Z
 thumbnail: /uploads/testtttt.jpg
+excerpt: This is the excerpt for the third post
 ---
+
 This is the third post
