+++ src/posts/second_post.md
@@ -2,6 +2,7 @@
 title: Second Post
 date: 2022-04-19T10:50:03.906Z
 thumbnail: /uploads/testtttt.jpg
+excerpt: This is the excerpt for the second post
 ---
 
 This is the second post
