+++ src/routes/blog/[slug].svelte
@@ -1,3 +1,15 @@
+<script context="module">
+	export async function load({ params }) {
+		const post = await import(`../../posts/${params.slug}.md`)
+		return {
+			props: {
+				postContent: post.default,
+				meta: post.metadata
+			}
+		}
+	}
+</script>
+
 <script>
 	export let postContent, meta
 	const { title, date, thumbnail } = meta
@@ -6,4 +18,5 @@
 <h1>{title}</h1>
 <p>{date}</p>
 <img src={thumbnail} alt={title} />
-<p>{postContent}</p>
+
+<svelte:component this={postContent} />
