+++ static/admin/config.yml
@@ -20,5 +20,6 @@ collections:
     fields:
       - { label: 'Title', name: 'title', widget: 'string', required: true }
       - { label: 'Publish Date', name: 'date', widget: 'datetime', required: true }
-      - { label: 'Featured Image', name: 'thumbnail', widget: 'image', required: false }
       - { label: 'Body', name: 'body', widget: 'markdown', required: true }
+      - { label: 'Excerpt', name: 'excerpt', widget: 'string', required: true }
+      - { label: 'Featured Image', name: 'thumbnail', widget: 'image', required: false }
